pyproject.toml
[tool.poetry]
name = "review-skill-django"
version = "0.1.0"
description = "ReviewSkill APIs"
authors = ["apella1 <johncollons5@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
black = "^23.10.1"
ruff = "^0.1.3"
django-on-heroku = "^1.1.2"
gunicorn = "^21.2.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
